requirements_frozen.txt
atlassian-python-api==1.14.8
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
docutils==0.15.2
flit==1.3
idna==2.8
importlib-metadata==1.3.0
intreehooks==1.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
more-itertools==8.0.2
mypy==0.761
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.3.2
pytoml==0.1.21
requests==2.22.0
requests-oauthlib==1.3.0
setuptools-scm==3.3.3
six==1.13.0
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.8
zipp==0.6.0
